requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pandas==2.2.3
scipy==1.14.1
groq==1.1.2
pydantic[email]==2.9.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
slowapi==0.1.9
psycopg2-binary==2.9.9
gunicorn==21.2.0
